requirements.txt
beautifulsoup4==4.6.0
Django==1.11.5
requests==2.18.4
djangorestframework==3.6.4
gunicorn
psycopg2
dj-database-url
whitenoise
